requirements.txt
django
flake8
django-polymorphic
